requirements.txt
django==5.1.6
psycopg2-binary==2.9.10
Pillow==11.1.0
requests==2.32.3
python-dotenv==1.0.1
whitenoise==6.9.0
gunicorn==23.0.0

# ML
transformers==4.49.0
torch==2.6.0
torchvision==0.21.0

# Supabase storage client
supabase==2.13.0
